fortune_handler_function/requirements.txt
azure-functions
requests~=2.24.0